Set Clamps
I did it! I have added about 10 lines of code from 90ish that specify boundaries for parameters going into both fitmixture and fitcircle. These raw parameter values, which are drawn from a gaussian, are clamped by the boundaries and an absolute distance for each parameter is summed and applied to the LL to assist with optimisation within fminsearch.