Fixed Lowerbound issue
There was a line in the most recent version of fitmixture4x.m and fitdcircle4x.m where a penalty was applied when eta, a or Ter was below zero. We made the decision to allow Ter to go negative, but did not reflect this in the boundary part. The tricky part was that it was further down in the code than the actual setting the parameter boundaries (Ub and Pub etc.)