EXPSIM
This is a new experiment where word and location stimuli are presented simultaneously. Most of the work done up to this point is in octobox, which is my attempt to change how the word is positioned relative to the circle, depending on which sector of the circle it is in. This is to avoid word clipping with the circle itself.