Checked fits of cdiffusion and criterionx
The .mat files labelled as the most recent fits did not match the table of fit statistics in draft 5 of the sourcemem manuscript. I have refit the models this afternoon, as I was unable to locate the .mat file I was looking for in the git commit history. Perhaps I was confused when putting the table together?
Should be good to proceed with results now.