Add joint fits between confidence ratings
In response to reviews, a joint fit of the low and high ratings of recognition for each of the models will be produced. A new folder has been made to keep this work distinct from the messiness of the repository to this point, which I will have to make sensible later.