Fix unintended constraint on mixture models in comparing across confidence levels
Unintentionally set a1 == a2 on the threhsold and hybrid diffusion models, as I wanted all parameters, besides mixture proportion/mean drift, to be the same across conditions. What 1 and 2 actually represent for a are the decision crtieria for memory and guessing processes, and these should be allowed to be different in this exercise.