Simulate imageability conditions separately
To avoid having to change the model fitting code at all from the version used for the manuscript fits, I have generated the exact number of observations in each imageability condition for each participant. In practice, this will not actually affect the data generated, since the parameters are set to be the same across imageability conditions, as we did not observe a difference between these conditions. Nonetheless, I am keeping things consistent for the purposes of model recovery.