Move the DRM lists into a junk folder
I dont want to do this in a DRM way because that approach maximises the similarity between each presented item and the unpresented critical lure. What we really want is to maximise the pairwise similarity of all the presented items. New idea is to take the word2vec embeddings and do a PCA on it, and then maybe the components can form the basis of lists?