De-merge experimental sessions
Figures out the revert messes. Turns out we do need separate templates for each session, because there is a line (400) that defines the data submission url, with an extension that tells the data handling what session the data belongs to. I can't figure out how to call the session number from within the html template, so separate templates seems like the easiest way around this.