Simulated response angle
Recentered errors should be calculated relative to the SIMULATED response angle (i.e. target angle + simulated error), and not the "response angle" column of the dataframe that is simply the observed data. the simulated data was matching the real one exactly, which is obvious now because im just plotting the data on top of itself.