Modify full model to reduce cond
Rewrote the full model code to be more flexible and allow for parameters to be equal across conditions (setting it to 0 will remove the penalty according to Sel, but will actually be interpreted as being equal to the first condition's value for that parameter). This allows for several steps of reduction to run on the same base model code.

Also implement a double fit so that each run of the model is fit twice, so that the estimated parameters of the first run are used as starting parameters for a second pass.