Strange eta
Trying to plot model predictions under different parameter conditions (showing that better response error fit results in worse response time fit), but eta is not behaving like I think it should. It is shifting the entire distribution to the right (so error is no longer centered on 0 but towards pi). Need to figure out why this is happening, maybe because eta1 is forced to = eta2