Free eta1 and eta2 to be different
So v2 should never be taken off 0, because we assume the target is canonically oriented, meaning it is always along the x axis (we just rotate the entire output around to make intrusion predictions), and v2 just shifts the entire distribution around by itself.

BUT, it does seem reasonable to have variation around v2, eta2, (i.e. variability in the tangential component of drift) separate from the variation in drift norm (eta1). When simulating the data (demo_diffusion_RT_misfit.m), this doesn't seem to make a huge difference in terms of fit, but need to see what happens when optimising without the constraint of eta1 = eta2. Have a look by running "run_spatiotemporal_eta" which takes the best fitting spatiotemporal model and finds separately values of eta1 and eta2 that might fit well.