eta1 != eta2
So, Klaus is right, constraining the two etas to be the same was hindering the model, but letting be different hasn't fixed everything (though it is a lot better).

I think it might be worth redoing some of the modelling just to make sure the story stays the same. Could be worth revisiting the previous publication as well. I think the constraint in the QQ is still the problem for the pure continuous model, but I want to be reassured.