Move response error model code
I want to rewrite the model function to take in a Pvar and Pfix vector and construct a parameter vector, which would allow me to run multiple models with the same basic function instead of having a billion separate functions for each variation of the model. I am moving legacy code to an old folder, so I can compare for equivalence with the rewritten function.