Replace upper with Pbounds[1,]
Upper specifies the upper boundary of estimated parameters. Taking the length of the upper vector tells us the number of free parameters to penalise in the AIC calculation. I have since replaced upper with Pbounds, so I can just take the length of one row of that for the same purpose.