Reparameterise and update response error model
The temporal weight being defined as 1-others is causing some problems when we want the temporal weight to be zero. Instead, we can assign parameters to each component, and normalise them. This does introduce an additional parameter, which is probably unnecessary. We could just write separate base model functions to work off, but that quickly gets confusing with the number of models we are working with here