Update response error model code
Main change here is I normalise the similarities (so after the exponential decay function, the resultant values are all divided by the maximum similarity). Previously, without this step, it was not possible to have steep drop off for, say, orthography, without lowering all the values of even the most orthographically similar item relative to the other types of similarity.

The models have been adjusted to account for this. Also, returned to nested wau of dealing with the different similarity factors, so that there are fewer parameters for the more complex models