Change how target is weighted
It just occurred to me I'm handling the target weighting in a dumb way, that sometimes lets it go negative for no good reason, and introduces a weird bounding to zero.

Instead, treat targets as a self-match, or an "intrusion" with a similarity weight of 1, no matter how the factors are balanced or whatever.